// web/src/api/generated/model/createCardBodyType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HouseApp API
 * API for HouseApp
 * OpenAPI spec version: 2.2.1
 */

export type CreateCardBodyType =
  (typeof CreateCardBodyType)[keyof typeof CreateCardBodyType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateCardBodyType = {
  physical: "physical",
  virtual: "virtual",
  additional: "additional",
} as const;